#336831: Refactor fix typo in package name
