+++ h1modules.utilities/src/net/sf/ecl1/utilities/standalone/workspace/FolderImpl.java
@@ -1,13 +1,16 @@
-package net.sf.ecl1.utilities.standalone.wokspace;
+package net.sf.ecl1.utilities.standalone.workspace;
 
+import java.io.File;
+import java.io.IOException;
 import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 import org.eclipse.core.resources.FileInfoMatcherDescription;
-import org.eclipse.core.resources.IBuildConfiguration;
 import org.eclipse.core.resources.IContainer;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IFolder;
@@ -15,7 +18,6 @@ import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.resources.IPathVariableManager;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IProjectDescription;
-import org.eclipse.core.resources.IProjectNature;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.resources.IResourceFilterDescription;
 import org.eclipse.core.resources.IResourceProxy;
@@ -26,223 +28,219 @@ import org.eclipse.core.resources.ResourceAttributes;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.QualifiedName;
-import org.eclipse.core.runtime.content.IContentTypeMatcher;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.ISchedulingRule;
 
-import net.sf.ecl1.utilities.hisinone.WebappsUtil;
 
-public class ProjectImpl implements IProject{
+public class FolderImpl implements IFolder {
 
-    private final String name;
-    private final Path projectPath;
+    private final Path path;
 
-    public ProjectImpl(String name, IPath workspacePath) {
-        this.name = name;
-        this.projectPath = Paths.get(workspacePath.toOSString(), name);
+    public FolderImpl(String path) {
+        this.path = Paths.get(path);
     }
 
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public IFolder getFolder(String name) {
-        return new FolderImpl(projectPath.resolve(name));
-    }
-
-    @Override
-    public IPath getLocation() {
-        return new PathImpl(projectPath);
-    }
-
-    @Override
-    public boolean exists(IPath path) {
-        return Files.exists(projectPath.resolve(path.toString()));
+    public FolderImpl(Path path) {
+        this.path = path;
     }
 
     @Override
     public boolean exists() {
-        return Files.exists(projectPath);
+        File folder = path.toFile();
+        return (folder.exists() && folder.isDirectory());
     }
 
     @Override
-    public IWorkspace getWorkspace() {
-        // always returns default workspace
-        return new WorkspaceImpl();
+    public boolean exists(IPath path) {
+        File folder = path.toFile();
+        return (folder.exists() && folder.isDirectory());
     }
 
     @Override
-    public int getType() {
-        return WorkspaceImpl.TYPE_PROJECT;
+    public IPath getLocation() {
+        return new PathImpl(path);
     }
 
     @Override
-    public IFile getFile(IPath path) {
-        return new FileImpl(projectPath.resolve(path.toString()));
+    public IPath getRawLocation() {
+       return getLocation();
     }
 
     @Override
-    public IFile getFile(String name) {
-        return new FileImpl(projectPath.resolve(name));
+    public IContainer getParent() {
+        String workspace = new WorkspaceRootImpl().getLocation().toString();
+        String parent = path.getParent().toString();
+        if(!parent.equals(workspace)){
+            return new FolderImpl(parent);
+        }
+        // parent is workspace root
+        return null;
     }
 
     @Override
-    public IFolder getFolder(IPath path) {
-        return new FolderImpl(projectPath.resolve(path.toString()));
+    public void create(boolean force, boolean local, IProgressMonitor monitor) throws CoreException {
+        try {
+            Files.createDirectories(path);
+        } catch (IOException e) {
+            throw new CoreException(new Status(IStatus.ERROR, FolderImpl.class, "Error creating directories at: " + path.toString()));
+        } 
+    }
+
+    @Override
+    public void create(int updateFlags, boolean local, IProgressMonitor monitor) throws CoreException {
+       create(false, local, monitor);
     }
 
     @Override
     public IPath getFullPath() {
-        return getFullPath(projectPath.toString());
+        return ProjectImpl.getFullPath(path.toString());
     }
 
-    protected static IPath getFullPath(String path){
-        Path workspacePath = Paths.get(new WorkspaceRootImpl().getLocation().toString());
-        Path projectPath = Paths.get(path);
-        // standalone only knows about projects in workspace-folder, handle webapps manually
-        IProject webapps = WebappsUtil.findWebappsProject();
-        String webappsName;
-        if(webapps != null){
-            webappsName = webapps.getName();
-            if(path.contains(webappsName)){
-                while(!projectPath.startsWith(webappsName)){
-                    projectPath = projectPath.subpath(1, projectPath.getNameCount());
+    @Override
+    public IResource[] members() throws CoreException {
+        File folder = path.toFile();
+        List<IResource> children = new ArrayList<>();
+        if (folder.exists() && folder.isDirectory()) {
+            // Get all files and subdirectories
+            File[] files = folder.listFiles();
+            if (files != null) {
+                for (File file : files) {
+                    if (file.isDirectory()) {
+                        children.add(new FolderImpl(file.getAbsolutePath()));
+                    } else {
+                        children.add(new FileImpl(file.getAbsolutePath()));
+                    }
                 }
-                return new PathImpl(projectPath.toString());
             }
         }
-        // use subpath to remove workspaceroot from path
-        Path relativePath = workspacePath.relativize(projectPath.subpath(1, projectPath.getNameCount())); 
-        return new PathImpl(relativePath.toString()); 
+        return children.toArray(new IResource[0]);
     }
 
     @Override
-    public void create(IProjectDescription description, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'create(IProjectDescription description, IProgressMonitor monitor)'");
+    public int getType() {
+        // cant get type for project
+        if(path.toString().equals(new WorkspaceRootImpl().getLocation().toString())){
+            return WorkspaceImpl.TYPE_ROOT;
+        }
+        File folder = path.toFile();
+        if(folder.exists() && folder.isDirectory()){
+            return WorkspaceImpl.TYPE_FOLDER;
+        }else if(folder.exists() && folder.isFile()){
+            return WorkspaceImpl.TYPE_FILE;
+        }
+        // no type
+        return 0;
     }
 
-    @Override
-    public void open(IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'open(IProgressMonitor monitor)'");
-    }
 
-    @Override
-    public void open(int updateFlags, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'open(int updateFlags, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void create(IProjectDescription description, int updateFlags, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'create(IProjectDescription description, int updateFlags, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void create(IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'create(IProgressMonitor monitor)'");
-    }
-    
     @Override
     public IResource findMember(String path) {
         throw new UnsupportedOperationException("Unimplemented method 'findMember(String path)'");
     }
-    
+
     @Override
     public IResource findMember(String path, boolean includePhantoms) {
         throw new UnsupportedOperationException("Unimplemented method 'findMember(String path, boolean includePhantoms)'");
     }
-    
+
     @Override
     public IResource findMember(IPath path) {
         throw new UnsupportedOperationException("Unimplemented method 'findMember(IPath path)'");
     }
-    
+
     @Override
     public IResource findMember(IPath path, boolean includePhantoms) {
         throw new UnsupportedOperationException("Unimplemented method 'findMember(IPath path, boolean includePhantoms)'");
     }
-    
+
     @Override
     public String getDefaultCharset() throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'getDefaultCharset()'");
     }
-    
+
     @Override
     public String getDefaultCharset(boolean checkImplicit) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'getDefaultCharset(boolean checkImplicit)'");
     }
-    
+
     @Override
-    public IResource[] members() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'members()'");
+    public IFile getFile(IPath path) {
+        throw new UnsupportedOperationException("Unimplemented method 'getFile(IPath path)'");
     }
-    
+
+    @Override
+    public IFolder getFolder(IPath path) {
+        throw new UnsupportedOperationException("Unimplemented method 'getFolder(IPath path)'");
+    }
+
     @Override
     public IResource[] members(boolean includePhantoms) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'members(boolean includePhantoms)'");
     }
-    
+
     @Override
     public IResource[] members(int memberFlags) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'members(int memberFlags)'");
     }
-    
+
     @Override
     public IFile[] findDeletedMembersWithHistory(int depth, IProgressMonitor monitor) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'findDeletedMembersWithHistory(int depth, IProgressMonitor monitor)'");
     }
-    
-    @Override
-    public void setDefaultCharset(String charset) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'setDefaultCharset(String charset)'");
-    }
-    
-    @Override
-    public void setDefaultCharset(String charset, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'setDefaultCharset(String charset, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public IResourceFilterDescription createFilter(int type, FileInfoMatcherDescription matcherDescription, int updateFlags, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'createFilter(int type, FileInfoMatcherDescription matcherDescription, int updateFlags, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public IResourceFilterDescription[] getFilters() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getFilters()'");
-    }
-    
-    @Override
-    public void accept(IResourceProxyVisitor visitor, int memberFlags) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceProxyVisitor visitor, int memberFlags)'");
-    }
-    
-    @Override
-    public void accept(IResourceProxyVisitor visitor, int depth, int memberFlags) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceProxyVisitor visitor, int depth, int memberFlags)'");
-    }
-    
-    @Override
-    public void accept(IResourceVisitor visitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor)'");
-    }
-    
-    @Override
-    public void accept(IResourceVisitor visitor, int depth, boolean includePhantoms) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor, int depth, boolean includePhantoms)'");
-    }
-    
-    @Override
-    public void accept(IResourceVisitor visitor, int depth, int memberFlags) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor, int depth, int memberFlags)'");
-    }
-    
+
     @Override
     public void clearHistory(IProgressMonitor monitor) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'clearHistory(IProgressMonitor monitor)'");
     }
-    
+
+    @Override
+    public void setDefaultCharset(String charset) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'setDefaultCharset(String charset)'");
+    }
+
+    @Override
+    public void setDefaultCharset(String charset, IProgressMonitor monitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'setDefaultCharset(String charset, IProgressMonitor monitor)'");
+    }
+
+    @Override
+    public IResourceFilterDescription createFilter(int type, FileInfoMatcherDescription matcherDescription,
+            int updateFlags, IProgressMonitor monitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'createFilter(int type, FileInfoMatcherDescription matcherDescription, int updateFlags, IProgressMonitor monitor)'");
+    }
+
+    @Override
+    public IResourceFilterDescription[] getFilters() throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'getFilters()'");
+    }
+
+    @Override
+    public void accept(IResourceProxyVisitor visitor, int memberFlags) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceProxyVisitor visitor, int memberFlags)'");
+    }
+
+    @Override
+    public void accept(IResourceProxyVisitor visitor, int depth, int memberFlags) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceProxyVisitor visitor, int depth, int memberFlags)'");
+    }
+
+    @Override
+    public void accept(IResourceVisitor visitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor)'");
+    }
+
+    @Override
+    public void accept(IResourceVisitor visitor, int depth, boolean includePhantoms) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor, int depth, boolean includePhantoms)'");
+    }
+
+    @Override
+    public void accept(IResourceVisitor visitor, int depth, int memberFlags) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'accept(IResourceVisitor visitor, int depth, int memberFlags)'");
+    }
+
     @Override
     public void copy(IPath destination, boolean force, IProgressMonitor monitor) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'copy(IPath destination, boolean force, IProgressMonitor monitor)'");
@@ -329,13 +327,13 @@ public class ProjectImpl implements IProject{
     }
     
     @Override
-    public IPathVariableManager getPathVariableManager() {
-        throw new UnsupportedOperationException("Unimplemented method 'getPathVariableManager()'");
+    public String getName() {
+        throw new UnsupportedOperationException("Unimplemented method 'getName()'");
     }
     
     @Override
-    public IContainer getParent() {
-        throw new UnsupportedOperationException("Unimplemented method 'getParent()'");
+    public IPathVariableManager getPathVariableManager() {
+        throw new UnsupportedOperationException("Unimplemented method 'getPathVariableManager()'");
     }
     
     @Override
@@ -359,11 +357,6 @@ public class ProjectImpl implements IProject{
     }
     
     @Override
-    public IPath getRawLocation() {
-        throw new UnsupportedOperationException("Unimplemented method 'getRawLocation()'");
-    }
-    
-    @Override
     public URI getRawLocationURI() {
         throw new UnsupportedOperationException("Unimplemented method 'getRawLocationURI()'");
     }
@@ -381,6 +374,11 @@ public class ProjectImpl implements IProject{
     @Override
     public Object getSessionProperty(QualifiedName key) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'getSessionProperty(QualifiedName key)'");
+    }
+    
+    @Override
+    public IWorkspace getWorkspace() {
+        throw new UnsupportedOperationException("Unimplemented method 'getWorkspace()'");
     }
     
     @Override
@@ -432,7 +430,7 @@ public class ProjectImpl implements IProject{
     public boolean isPhantom() {
         throw new UnsupportedOperationException("Unimplemented method 'isPhantom()'");
     }
-
+    
     @Override
     public boolean isReadOnly() {
         throw new UnsupportedOperationException("Unimplemented method 'isReadOnly()'");
@@ -452,7 +450,7 @@ public class ProjectImpl implements IProject{
     public boolean isTeamPrivateMember(int options) {
         throw new UnsupportedOperationException("Unimplemented method 'isTeamPrivateMember(int options)'");
     }
-    
+
     @Override
     public void move(IPath destination, boolean force, IProgressMonitor monitor) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'move(IPath destination, boolean force, IProgressMonitor monitor)'");
@@ -512,8 +510,8 @@ public class ProjectImpl implements IProject{
     @Override
     public void setPersistentProperty(QualifiedName key, String value) throws CoreException {
         throw new UnsupportedOperationException("Unimplemented method 'setPersistentProperty(QualifiedName key, String value)'");
-    }    
-
+    }
+    
     @Override
     public void setReadOnly(boolean readOnly) {
         throw new UnsupportedOperationException("Unimplemented method 'setReadOnly(boolean readOnly)'");
@@ -555,135 +553,33 @@ public class ProjectImpl implements IProject{
     }
     
     @Override
-    public void build(int kind, String builderName, Map<String, String> args, IProgressMonitor monitor)
+    public void createLink(IPath localLocation, int updateFlags, IProgressMonitor monitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'createLink(IPath localLocation, int updateFlags, IProgressMonitor monitor)'");
+    }
+    
+    @Override
+    public void createLink(URI location, int updateFlags, IProgressMonitor monitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'createLink(URI location, int updateFlags, IProgressMonitor monitor)'");
+    }
+    
+    @Override
+    public void delete(boolean force, boolean keepHistory, IProgressMonitor monitor) throws CoreException {
+        throw new UnsupportedOperationException("Unimplemented method 'delete(boolean force, boolean keepHistory, IProgressMonitor monitor)'");
+    }
+    
+    @Override
+    public IFile getFile(String name) {
+        throw new UnsupportedOperationException("Unimplemented method 'getFile(String name)'");
+    }
+    
+    @Override
+    public IFolder getFolder(String name) {
+        throw new UnsupportedOperationException("Unimplemented method 'getFolder(String name)'");
+    }
+    
+    @Override
+    public void move(IPath destination, boolean force, boolean keepHistory, IProgressMonitor monitor)
             throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'build(int kind, String builderName, Map<String, String> args, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void build(int kind, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'build(int kind, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void build(IBuildConfiguration config, int kind, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'build(IBuildConfiguration config, int kind, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void close(IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'close(IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void delete(boolean deleteContent, boolean force, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'delete(boolean deleteContent, boolean force, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public IBuildConfiguration getActiveBuildConfig() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getActiveBuildConfig()'");
-    }
-    
-    @Override
-    public IBuildConfiguration getBuildConfig(String configName) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getBuildConfig(String configName)'");
-    }
-    
-    @Override
-    public IBuildConfiguration[] getBuildConfigs() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getBuildConfigs()'");
-    }
-    
-    @Override
-    public IContentTypeMatcher getContentTypeMatcher() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getContentTypeMatcher()'");
-    }
-    
-    @Override
-    public IProjectDescription getDescription() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getDescription()'");
-    }
-    
-    @Override
-    public IProjectNature getNature(String natureId) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getNature(String natureId)'");
-    }
-    
-    @Override
-    public IPath getWorkingLocation(String id) {
-        throw new UnsupportedOperationException("Unimplemented method 'getWorkingLocation(String id)'");
-    }
-    
-    @Override
-    public IProject[] getReferencedProjects() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getReferencedProjects()'");
-    }
-    
-    @Override
-    public void clearCachedDynamicReferences() {
-        throw new UnsupportedOperationException("Unimplemented method 'clearCachedDynamicReferences()'");
-    }
-    
-    @Override
-    public IProject[] getReferencingProjects() {
-        throw new UnsupportedOperationException("Unimplemented method 'getReferencingProjects()'");
-    }
-    
-    @Override
-    public IBuildConfiguration[] getReferencedBuildConfigs(String configName, boolean includeMissing)
-            throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getReferencedBuildConfigs(String configName, boolean includeMissing)'");
-    }
-    
-    @Override
-    public boolean hasBuildConfig(String configName) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'hasBuildConfig(String configName)'");
-    }
-    
-    @Override
-    public boolean hasNature(String natureId) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'hasNature(String natureId)'");
-    }
-    
-    @Override
-    public boolean isNatureEnabled(String natureId) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'isNatureEnabled(String natureId)'");
-    }
-    
-    @Override
-    public boolean isOpen() {
-        throw new UnsupportedOperationException("Unimplemented method 'isOpen()'");
-    }
-    
-    @Override
-    public void loadSnapshot(int options, URI snapshotLocation, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'loadSnapshot(int options, URI snapshotLocation, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void move(IProjectDescription description, boolean force, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'move(IProjectDescription description, boolean force, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void saveSnapshot(int options, URI snapshotLocation, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'saveSnapshot(int options, URI snapshotLocation, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void setDescription(IProjectDescription description, IProgressMonitor monitor) throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'setDescription(IProjectDescription description, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public void setDescription(IProjectDescription description, int updateFlags, IProgressMonitor monitor)
-            throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'setDescription(IProjectDescription description, int updateFlags, IProgressMonitor monitor)'");
-    }
-    
-    @Override
-    public String getDefaultLineSeparator() throws CoreException {
-        throw new UnsupportedOperationException("Unimplemented method 'getDefaultLineSeparator()'");
+        throw new UnsupportedOperationException("Unimplemented method 'move(IPath destination, boolean force, boolean keepHistory, IProgressMonitor monitor)'");
     }
 }
