+++ h1modules.utilities/src/net/sf/ecl1/utilities/standalone/workspace/WorkspaceRootImpl.java
@@ -1,4 +1,4 @@
-package net.sf.ecl1.utilities.standalone.wokspace;
+package net.sf.ecl1.utilities.standalone.workspace;
 
 import java.io.File;
 import java.net.URI;
