+++ h1modules/src/de/his/cs/sys/extensions/standalone/NewExtensionWizardPageStandalone.java
@@ -26,7 +26,7 @@ import net.sf.ecl1.utilities.general.InitialProjectConfigurationChoices;
 import net.sf.ecl1.utilities.general.WorkspaceSupport;
 import net.sf.ecl1.utilities.hisinone.HisConstants;
 import net.sf.ecl1.utilities.hisinone.WebappsUtil;
-import net.sf.ecl1.utilities.standalone.wokspace.WorkspaceFactory;
+import net.sf.ecl1.utilities.standalone.workspace.WorkspaceFactory;
 
 /**
  * Extended New Project Wizard Page asking for additional information on extensions
