+++ net.sf.ecl1.import.wizard/src/net/sf/ecl1/importwizard/ExtensionImportJob.java
@@ -20,7 +20,7 @@ import com.google.common.collect.Lists;
 import net.sf.ecl1.utilities.logging.ICommonLogger;
 import net.sf.ecl1.utilities.logging.LoggerFactory;
 import net.sf.ecl1.utilities.preferences.PreferenceWrapper;
-import net.sf.ecl1.utilities.standalone.wokspace.WorkspaceFactory;
+import net.sf.ecl1.utilities.standalone.workspace.WorkspaceFactory;
 
 /**
  * The job that actually performs deleting of existing folders in the workspace and the extension import.
