+++ net.sf.ecl1.import.wizard/src/net/sf/ecl1/importwizard/ExtensionImportWizard.java
@@ -11,7 +11,7 @@ import org.eclipse.jface.wizard.Wizard;
 import org.eclipse.ui.IImportWizard;
 import org.eclipse.ui.IWorkbench;
 
-import net.sf.ecl1.utilities.standalone.wokspace.WorkspaceFactory;
+import net.sf.ecl1.utilities.standalone.workspace.WorkspaceFactory;
 
 /**
  * Wizard for import of extension projects listed on a jenkins.
